Fix sorting order of mounts
Summary:
--------
GNU sort uses collation order of current locale by default, which was
previously fixed by a workaround. This commit uses:
LC_ALL=C sort ...
where it must be used and in one case where it may matter in the future.

Long version:
-------------
This one need some explanation, because it worked due to a workaround.

To mount everything in correct order, the installer sorts its config by
mountpoint using an internal function FSspec_sort().

A) Workaround
But it calls this function to sort in reverse order - which would be
wrong for mounting.

Example(simplified): desired result from sorting:
 sda3:swap
/ sda2:ext4
/boot sda1:ext2

The above is what is returned from FSspec_sort(), but why with reverse
order?

B) The real problem
The real problem is within that FSspec_sort() function.
It works fine when sorting by partition, but when sorting by mountpoint,
it requires reversed sorting.
Internally it uses 'sort', which has a gotcha - it uses collation order
of the current locale:
https://www.gnu.org/software/coreutils/faq/coreutils-faq.html#Sort-does-not-sort-in-normal-order_0021

Example (without fix): Run this code:
echo ' sda3:swap
/ sda2:ext4
/boot sda1:ext2' | sort

Result:
/boot sda1:ext2
/ sda2:ext4
 sda3:swap

This is the wrong order for mounting ... but fixed by the workaround
'sort reverse'.

Example (with fix):
echo ' sda3:swap
/ sda2:ext4
/boot sda1:ext2' | LC_ALL=C sort

Result:
 sda3:swap
 / sda2:ext4
 /boot sda1:ext2

This is what we want, no need for a workaround.

I think I just gave up on this when I wrote this years ago and decided
to live with the workaround.