Move chroot-emerge before unmounting
These 2 lines did not check exit status for a reason, they returned
errors. From the logs:

+ chroot /mnt/gentoo emerge --noreplace sys-boot/grub:2 --nodeps
Failed to validate a sane '/dev'.
bash process substitution doesn't work; this may be an indication of a
broken '/dev/fd'.
+ chroot /mnt/gentoo emerge --rage-clean sys-boot/grub:0
Failed to validate a sane '/dev'.
bash process substitution doesn't work; this may be an indication of a
broken '/dev/fd'.

Reason for this was that these lines were after unmounting target
partitions.