feature: Unsafe skip verify
This adds a flag for `--unsafe-skip-verify` allowing users who have not
booted with the verify option to skip this while forcing them to
acknowledge the inherant concerns with skipping digest checking of
resources.

Additionally, this adds the variable `INSTALLER_UNSAFE` which allows for
determining if the verification step has been skipped.