eliminate uname -r usage
There simply isn't a need to be this selective with things.  As livecd
and installer use cases grow, we could be booted into a non-pentoo
kernel and installing via a chroot (yeah, I've done it).  So let's just
skip the error and sync the kernel modules and call it a day.